Increase logging, record each command as run
Add a new function that records a command into the log file, then runs
the command.  Make use of this command throughout the build-all.sh
script so that the log file contains what the command is that gets run,
then the log of the command being run.  In this way it should be easier
to reproduce any errors without having to read the build-all.sh script.