build-all.sh: Rework the configure options for binutils and gcc
I've gone over the configure options for binutils-gdb and gcc, and
removed some of the contradictory flags that were present.

I've also moved the value for --with-cpu into a variable in the script
so it is controlled from one place.

The only thing that I am unhappy about is the use of
--disable-multilib.  This is how mellanox configures their tools in
buildroot, but it would be nice if I did not do that in this script (so
I could test all targets).  However, the stage-2 GCC currently does not
build with multilibs turned on.  I think the problem might be related
to either --enable-static or --enable-shared, but it needs more
investigation.